fix: alarm crash on Android 12+ and next batch of Kotlin warnings
AlarmService registered for call-state changes with
TelephonyManager.listen(LISTEN_CALL_STATE). Apps targeting S+ must hold
READ_PHONE_STATE for that (the app does not declare it), so it throws
SecurityException as soon as an alarm starts ringing. Now:
- API 31+: TelephonyCallback.CallStateListener, registered only when
  READ_PHONE_STATE is granted; otherwise the "mark missed on incoming
  call" behavior is skipped instead of crashing.
- API 29/30: the legacy PhoneStateListener (no permission needed there),
  with the deprecation suppressed on that version-gated path only.
- Any SecurityException while registering is caught and logged.
- stopForeground(Boolean) -> stopForeground(STOP_FOREGROUND_REMOVE).

Other warnings from the latest APK build: redundant Int conversion in
AlarmStateManager, unnecessary !! in AlarmItemViewHolder, and the
setTargetFragment call androidx.preference still requires is now
suppressed at that line.

CI: Gradle only prints the first ~15 compiler problems to the console.
New scripts/ci/print_compiler_warnings.py reads the full list from
build/reports/problems/problems-report.html and prints every one (with
GitHub annotations) after the APK build.

+++ app/src/main/kotlin/com/android/deskclock/alarms/AlarmStateManager.kt
@@ -334,7 +334,7 @@ class AlarmStateManager : BroadcastReceiver() {
             intent.addCategory(tag)
             intent.putExtra(ALARM_GLOBAL_ID_EXTRA, DataModel.dataModel.globalIntentId)
             if (state != null) {
-                intent.putExtra(ALARM_STATE_EXTRA, state.toInt())
+                intent.putExtra(ALARM_STATE_EXTRA, state)
             }
             return intent
         }
